refactor: ChatsHome and componenting chats background
